refactor: hooks
